MFCC and AddNoise
In extractMFCC there are two functions to compute the features, one takes in input a name of an audio file, the other takes in input a vector representing an audio file.
addNoise has a function named addNoise that adds noise to a signal. Intensity of the noise to add is a parameter. It takes in input the noise audio file name.
I upload also the audio tracks used to test the code.